Actually add containing tests to Tera and move tests around a bit
