+++ src/tera.rs
@@ -496,6 +496,7 @@ impl Tera {
         self.register_tester("iterable", testers::iterable);
         self.register_tester("starting_with", testers::starting_with);
         self.register_tester("ending_with", testers::ending_with);
+        self.register_tester("containing", testers::containing);
     }
 
     fn register_tera_global_functions(&mut self) {
